Double-encode entities for turndown
Turndown uses the raw text value to further output markdown, which will
result in escaped input like `&lt;node foo="bar" /&gt;` to be output
like `<node foo="bar" />` and in turn treated as inline HTML.
